refactor(gateway): preserve one serving protocol
